+++ RobWork/src/rw/invkin/ParallelIKSolver.cpp
@@ -64,7 +64,7 @@ std::vector<Q> ParallelIKSolver::solve(const Transform3D<>& dest,
     std::vector< ParallelLeg* > legs = _device->getLegs();
 
     // initialize currQ
-    Q currQ(Q::zero(legs.size() * 6));
+    Q currQ(Q::zero((int)legs.size() * 6));
 
     size_t columns=0, qIndex;
     std::vector<ParallelLeg*>::iterator leg_iter = legs.begin();
@@ -81,10 +81,10 @@ std::vector<Q> ParallelIKSolver::solve(const Transform3D<>& dest,
     }
 
     // initialize deltaQ
-    Q deltaQ(Q::zero(legs.size()*6));
+    Q deltaQ(Q::zero((int)legs.size()*6));
 
     // initialize deltaX vector
-    Q deltaX(Q::zero(legs.size()*6));
+    Q deltaX(Q::zero((int)legs.size()*6));
 
     // calculate the difference from current pose to dest pose
     Transform3D<double> curr = legs.front()->baseTend(state);
@@ -136,7 +136,7 @@ std::vector<Q> ParallelIKSolver::solve(const Transform3D<>& dest,
                 Vector3D<double> pos = bTe_1.P() - bTe.P();
                 EAA<double> orin = bTe.R()*(EAA<>( inverse(bTe.R())*bTe_1.R() ) );
                 // copy it into deltaX
-                int index = i*6;
+                int index = (int)i*6;
                 deltaX[index+0] = pos(0);
                 deltaX[index+1] = pos(1);
                 deltaX[index+2] = pos(2);
@@ -205,7 +205,7 @@ std::vector<Q> ParallelIKSolver::solve(const Transform3D<>& dest,
                 Vector3D<double> pos = bTe_1.P() - bTe.P();
                 EAA<double> orin = bTe.R()*(EAA<>( inverse(bTe.R())*bTe_1.R() ) );
                 // copy it into deltaX
-                int index = i*6;
+                int index = (int)i*6;
                 deltaX[index+0] = pos(0);
                 deltaX[index+1] = pos(1);
                 deltaX[index+2] = pos(2);
